ComputationalGraphs

Documentation for ComputationalGraphs.

ComputationalGraphs.ComputationalGraph

ComputationalGraphs.add_dependency!

ComputationalGraphs.add_node!

ComputationalGraphs.compute_all_nodes!

ComputationalGraphs.compute_computable_nodes!

ComputationalGraphs.compute_computable_nodes_pmap!

ComputationalGraphs.compute_node!

ComputationalGraphs.compute_node_data

ComputationalGraphs.get_computable_nodes

ComputationalGraphs.is_computable

ComputationalGraphs.is_computed

ComputationalGraphs.reachable_nodes

ComputationalGraphs.unassign_intermediate_nodes!

ComputationalGraphs.update_node_value!

ComputationalGraphs.ComputationalGraph — Type

A simple computational graph implementation

ComputationalGraphs.add_dependency! — Method

Add a dependency to a node in the computational graph.

ComputationalGraphs.add_node! — Method

Add a node to the computational graph.

ComputationalGraphs.compute_all_nodes! — Method

Compute all nodes in the computational graph. Return the number of calls of computecomputablenodes!.

ComputationalGraphs.compute_computable_nodes! — Method

Compute all computable nodes in the computational graph at the time of calling. Returns the number of computed nodes.

You can call this function multiple times, and it will only compute nodes that are computable at the time of calling.

ComputationalGraphs.compute_computable_nodes_pmap! — Method

Like computecomputablenodes!, but computes nodes in parallel using distributed workers. Returns the number of computed nodes.

ComputationalGraphs.compute_node! — Method

Compute the value of a node in the computational graph.

ComputationalGraphs.compute_node_data — Method

Prepare minimal data (dependencies' values + computefunc) for a node.

ComputationalGraphs.get_computable_nodes — Method

Get the computable nodes in the computational graph. Ignore nodes that are already computed or assigned a value.

ComputationalGraphs.is_computable — Method

Return if a node is computable, i.e., all its dependencies have been computed.

ComputationalGraphs.is_computed — Method

Return if a node is computed, i.e., it has a value assigned.

ComputationalGraphs.reachable_nodes — Method

Find all nodes reachable from a starting node in a directed graph using depth-first search.

Arguments

g::NamedDiGraph: The directed graph to search

start_node::Int: The node ID to start searching from

Returns

Vector{Int}: IDs of all nodes reachable from the start node

ComputationalGraphs.unassign_intermediate_nodes! — Method

Unassigned all unneseccary intermediate nodes from the computational graph. Return the number of unassigned nodes.

ComputationalGraphs.update_node_value! — Method

Update the value of a node and invalidate all dependent nodes.

Arguments

obj::ComputationalGraph: The computational graph

name::K: Name of the node to update

nodevalue::N: New value to assign to the node